faculty controller

+++ Service/Application/Services/FacultyService.cs
@@ -4,28 +4,35 @@ using Infrastructure.Context;
 using Infrastructure.Repositories;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.IdentityModel.Tokens;
+using XAct.Users;
 
 namespace Application.Services
 {
     public interface IFacultyService
     {
-        Tuple<string, Faculty?> Add(AddFacultyModel a);
-        Tuple<string, Faculty?> Update(EditFacultyModel e);
-        Tuple<string, bool> Delete(string id);
+        Tuple<string, Faculty?> Add(string userId, AddFacultyModel a);
+        Tuple<string, Faculty?> Update(string userId, EditFacultyModel e);
+        Tuple<string, bool> Delete(string userId, string id);
         Tuple<string, Faculty?> GetById(string id);
-        Tuple<string, List<Faculty?>?> GetAll();
+        Tuple<string, List<Faculty?>?> GetAll(string userId);
+        Tuple<string, List<Faculty?>?> GetPublic();
     }
     public class FacultyService:IFacultyService
     {
         private readonly IUnitOfWork unitOfWork;
+
         public FacultyService(UserDbContext context, IMemoryCache cache)
         {
             unitOfWork = new UnitOfWork(context, cache);
         }
 
-        public Tuple<string, Faculty?> Add(AddFacultyModel a)
+        public Tuple<string, Faculty?> Add(string userId, AddFacultyModel a)
         {
             if (a == null) return new Tuple<string, Faculty?>("parameter is null", null);
+            if (userId.IsNullOrEmpty()) return new Tuple<string, Faculty?>("userId not null", null);
+            var user = unitOfWork.userRepository.GetById(userId);
+            if (user == null) return new Tuple<string, Faculty?>($"not found with userId: {userId}", null);
+            if (user.Role.Name != "ADMIN") return new Tuple<string, Faculty?>("you cant do this", null);
             var newFaculty = new Faculty()
             {
                 Id = Guid.NewGuid().ToString(),
@@ -36,20 +43,29 @@ namespace Application.Services
             return new Tuple<string, Faculty?>("add successful!", newFaculty);
         }
 
-        public Tuple<string, bool> Delete(string id)
+        public Tuple<string, bool> Delete(string userId, string id)
         {
             if (id.IsNullOrEmpty())
             {
                 return new Tuple<string, bool>("parameter is null!", false);
             }
+            if (userId.IsNullOrEmpty()) return new Tuple<string, bool>("userId not null", false);
+            var user = unitOfWork.userRepository.GetById(userId);
+            if (user == null) return new Tuple<string, bool>($"not found with userId: {userId}", false);
+            if (user.Role.Name != "ADMIN") return new Tuple<string, bool>("you cant do this", false);
             var faculty = unitOfWork.facultyRepository.GetById(id);
+            if (faculty == null) return new Tuple<string, bool>($"not found faculty with id: {id}", false);
             unitOfWork.facultyRepository.Remove(faculty);
             unitOfWork.SaveChange();
             return new Tuple<string, bool>("delete successful", true);
         }
 
-        public Tuple<string, List<Faculty?>?> GetAll()
+        public Tuple<string, List<Faculty?>?> GetAll(string userId)
         {
+            if (userId.IsNullOrEmpty()) return new Tuple<string, List<Faculty?>?>("userId not null", null);
+            var user = unitOfWork.userRepository.GetById(userId);
+            if (user == null) return new Tuple<string, List<Faculty?>?>($"not found with userId: {userId}", null);
+            if (user.Role.Name != "ADMIN") return new Tuple<string, List<Faculty?>?>("you cant do this", null);
             var allFaculty = unitOfWork.facultyRepository.GetAll();
             if (allFaculty.IsNullOrEmpty())
             {
@@ -69,9 +85,23 @@ namespace Application.Services
             return new Tuple<string, Faculty?>("", faculty);
         }
 
-        public Tuple<string, Faculty?> Update(EditFacultyModel e)
+        public Tuple<string, List<Faculty?>?> GetPublic()
+        {
+            var publicfaculty = unitOfWork.facultyRepository.Find(f => f.IsDeleted == false);
+            if (publicfaculty.IsNullOrEmpty())
+            {
+                return new Tuple<string, List<Faculty?>?>("no faculty public", null);
+            }
+            return new Tuple<string, List<Faculty?>?>("", publicfaculty);
+        }
+
+        public Tuple<string, Faculty?> Update(string userId, EditFacultyModel e)
         {
             if (e == null) return new Tuple<string, Faculty?>("parameter is null", null);
+            if (userId.IsNullOrEmpty()) return new Tuple<string, Faculty?>("userId not null", null);
+            var user = unitOfWork.userRepository.GetById(userId);
+            if (user == null) return new Tuple<string, Faculty?>($"not found with userId: {userId}", null);
+            if (user.Role.Name != "ADMIN") return new Tuple<string, Faculty?>("you cant do this", null);
             var faculty = unitOfWork.facultyRepository.GetById(e.Id);
             if (faculty == null) return new Tuple<string, Faculty?>($"not founf with id: {e.Id}", null);
             faculty.Name = e.Name;
